move quadtree code to separate file
